attempting to make working server
