+++ Gemfile
@@ -1,7 +1,5 @@
 source 'http://rubygems.org'
 
-gem 'sinatra'
-
 gem 'magma-dev-ucsf-immunoprofiler'
-
+gem 'rack'
 gem 'pg'
