+++ config.ru
@@ -2,7 +2,7 @@ require_relative 'gem/lib/magma/server'
 require_relative 'gem/lib/magma'
 require 'yaml'
 
-use Magma::Auth
+use Magma::IpAuth
 use Magma::JsonBody
 
 run Magma::Server.new(YAML.load File.read("config.yml"))
